+++ philo/src/utils/philo_list_helpers/get_philo.c
@@ -6,13 +6,13 @@
 /*   By: pablo <pablo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/19 22:40:14 by pablo             #+#    #+#             */
-/*   Updated: 2025/06/19 22:55:22 by pablo            ###   ########.fr       */
+/*   Updated: 2025/06/26 12:17:19 by pablo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-t_philo	*get_philo(t_philo *philos, size_t id)
+t_philo	*get_philo(t_philo *philos, unsigned int id)
 {
 	t_philo	*start;
